Feature/fix indexing typecasting issue

+++ rvic/core/param_file.py
@@ -49,7 +49,7 @@ def finish_params(outlets, dom_data, config_dict, directories):
     # subset (shorten time base)
     if options['SUBSET_DAYS'] and \
             options['SUBSET_DAYS'] < routing['BASIN_FLOWDAYS']:
-        subset_length = (options['SUBSET_DAYS'] *
+        subset_length = int(options['SUBSET_DAYS'] *
                          SECSPERDAY / routing['OUTPUT_INTERVAL'])
         outlets, full_time_length, \
             before, after = subset(outlets, subset_length=subset_length)
@@ -330,8 +330,8 @@ def subset(outlets, subset_length=None):
         out_uh = np.zeros((subset_length, outlet.unit_hydrograph.shape[1]),
                           dtype=np.float64)
 
-        d_left = -1 * subset_length / 2
-        d_right = subset_length / 2
+        d_left = int(-1 * subset_length / 2)
+        d_right = int(subset_length / 2)
 
         for j in pyrange(outlet.unit_hydrograph.shape[1]):
             # find index position of maximum
